Refactoring and code cleaning due to circular import issues
