+++ src/main/go/api/rest/portfolio_allocation_controller.go
@@ -11,7 +11,8 @@ import (
 	"github.com/benizzio/open-asset-allocator/domain"
 	"github.com/benizzio/open-asset-allocator/domain/service"
 	"github.com/benizzio/open-asset-allocator/infra"
-	"github.com/benizzio/open-asset-allocator/infra/util"
+	gininfra "github.com/benizzio/open-asset-allocator/infra/gin"
+	"github.com/benizzio/open-asset-allocator/infra/validation"
 	"github.com/benizzio/open-asset-allocator/langext"
 	"github.com/gin-gonic/gin"
 )
@@ -152,7 +153,7 @@ func (controller *PortfolioAllocationRESTController) postPortfolioAllocationHist
 	}
 
 	var portfolioSnapshotDTS model.PortfolioSnapshotDTS
-	valid, err := util.BindAndValidateJSONWithInvalidResponse(context, &portfolioSnapshotDTS)
+	valid, err := gininfra.BindAndValidateJSONWithInvalidResponse(context, &portfolioSnapshotDTS)
 	if infra.HandleAPIError(context, bindPortfolioSnapshotErrorMessage, err) || !valid {
 		return
 	}
@@ -197,7 +198,7 @@ func (controller *PortfolioAllocationRESTController) validateCleanPortfolioAlloc
 	)
 	portfolioSnapshotDTS.Allocations = cleanAllocations
 
-	var validationErrorsBuilder = util.BuildCustomValidationErrorsBuilder()
+	var validationErrorsBuilder = validation.BuildCustomValidationErrorsBuilder()
 	for index, allocation := range portfolioSnapshotDTS.Allocations {
 
 		var isAssetIdentified = langext.IsZeroValue(allocation.AssetId) &&
@@ -215,7 +216,7 @@ func (controller *PortfolioAllocationRESTController) validateCleanPortfolioAlloc
 
 	var validationErrors = validationErrorsBuilder.Build()
 	if len(validationErrors) > 0 {
-		util.RespondWithCustomValidationErrors(context, validationErrors, portfolioSnapshotDTS)
+		gininfra.RespondWithCustomValidationErrors(context, validationErrors, portfolioSnapshotDTS)
 		return false
 	}
