+++ src/main/go/domain/hierarchical_id_test.go
@@ -2,27 +2,11 @@ package domain
 
 import (
 	"database/sql"
-	"database/sql/driver"
-	"fmt"
 	"testing"
 
+	"github.com/benizzio/open-asset-allocator/infra/util"
 	"github.com/lib/pq"
 )
-
-// valueToString normalizes driver.Value to a comparable string.
-// It handles string and []byte, and falls back to fmt.Sprintf otherwise.
-//
-// Authored by: GitHub Copilot
-func valueToString(v driver.Value) string {
-	switch x := v.(type) {
-	case string:
-		return x
-	case []byte:
-		return string(x)
-	default:
-		return fmt.Sprintf("%v", v)
-	}
-}
 
 func TestHierarchicalIdValue_AllNonNil(t *testing.T) {
 	var a = "a"
@@ -30,20 +14,24 @@ func TestHierarchicalIdValue_AllNonNil(t *testing.T) {
 	var c = "c"
 
 	var id = HierarchicalId{&a, &b, &c}
-	got, err := id.Value()
+	actual, err := id.Value()
 	if err != nil {
 		t.Fatalf("Value() returned error: %v", err)
 	}
 
 	// Expected representation using pq.Array on []sql.NullString
-	expectedArray := []sql.NullString{{String: "a", Valid: true}, {String: "b", Valid: true}, {String: "c", Valid: true}}
+	expectedArray := []sql.NullString{
+		{String: "a", Valid: true},
+		{String: "b", Valid: true},
+		{String: "c", Valid: true},
+	}
 	expected, err := pq.Array(expectedArray).Value()
 	if err != nil {
 		t.Fatalf("pq.Array.Value() returned error: %v", err)
 	}
 
-	if valueToString(got) != valueToString(expected) {
-		t.Fatalf("mismatch\n got: %q\nwant: %q", valueToString(got), valueToString(expected))
+	if util.ValueToString(actual) != util.ValueToString(expected) {
+		t.Fatalf("mismatch\n actual: %q\nwant: %q", util.ValueToString(actual), util.ValueToString(expected))
 	}
 }
 
@@ -52,18 +40,22 @@ func TestHierarchicalIdValue_WithNilLevels(t *testing.T) {
 	var c = "c"
 
 	var id = HierarchicalId{&a, nil, &c}
-	got, err := id.Value()
+	actual, err := id.Value()
 	if err != nil {
 		t.Fatalf("Value() returned error: %v", err)
 	}
 
-	expectedArray := []sql.NullString{{String: "a", Valid: true}, {String: "", Valid: false}, {String: "c", Valid: true}}
+	expectedArray := []sql.NullString{
+		{String: "a", Valid: true},
+		{String: "", Valid: false},
+		{String: "c", Valid: true},
+	}
 	expected, err := pq.Array(expectedArray).Value()
 	if err != nil {
 		t.Fatalf("pq.Array.Value() returned error: %v", err)
 	}
 
-	if valueToString(got) != valueToString(expected) {
-		t.Fatalf("mismatch with nils\n got: %q\nwant: %q", valueToString(got), valueToString(expected))
+	if util.ValueToString(actual) != util.ValueToString(expected) {
+		t.Fatalf("mismatch with nils\n actual: %q\nwant: %q", util.ValueToString(actual), util.ValueToString(expected))
 	}
 }
