+++ src/main/go/infra/util/sql.go
@@ -1,15 +1,34 @@
 package util
 
-import "database/sql"
+import (
+	"database/sql"
+	"database/sql/driver"
+	"fmt"
+)
 
-func ToNullString(str string) sql.NullString {
-	return ToNullStringFromPointer(&str)
+func StringToNullString(str string) sql.NullString {
+	return StringPointerToNullString(&str)
 }
 
-func ToNullStringFromPointer(str *string) sql.NullString {
+func StringPointerToNullString(str *string) sql.NullString {
 	if str == nil {
 		return sql.NullString{Valid: false}
 	} else {
 		return sql.NullString{String: *str, Valid: true}
+	}
+}
+
+// ValueToString normalizes driver.Value to a comparable string.
+// It handles string and []byte, and falls back to fmt.Sprintf otherwise.
+//
+// Co-authored by: GitHub Copilot
+func ValueToString(value driver.Value) string {
+	switch valueType := value.(type) {
+	case string:
+		return valueType
+	case []byte:
+		return string(valueType)
+	default:
+		return fmt.Sprintf("%v", value)
 	}
 }
